requirements.txt
Telethon==1.43.2
python-dotenv==1.2.2

